Give constant a name and lift color definition up
